拆分图表为单独组件
